Outpainting works with smaller steps, upscaling, retrained models

+++ clean_up_wandb.py
@@ -4,13 +4,24 @@ import wandb
 api = wandb.Api()
 
 
-for i in range(195):
-    #print(i)
+
+
+artifact_path = "matezzzz/image_outpainting_maskgit/run_xk236zct_model"
+
+keep_each = 20
+delete_from = 0
+delete_to = 57
+
+
+for i in range(delete_from, delete_to+1):
+    if i % keep_each == 0:
+        print ("Keeping", i)
+        continue
     try:
-        a = api.artifact(f'matezzzz/image_outpainting_tokenizer/run_2llyronh_model:v{i}')
+        a = api.artifact(f'{artifact_path}:v{i}')
         print ("Deleting", i)
     except wandb.errors.CommError:
         print ("Can't find", i)
-        continue    
+        continue
     a.delete(delete_aliases=True)
     a.delete()
